# reduce contention of global string cache: `>4x` performance improvement