This greatly reduces the contention of the global string cache, therefore greatly improving the performance of casting many columns.

The contention time has reduced from `O(n)` to `O(k)` where

* `n` is number of elements
* `k` is the number of unique elements

My local benchmark showed a **4x** improvement, but I think it will show much more improvement if you have real categorical data, e.g. where `k` is much lower than `n`.

```python
>>> df = pl.read_csv("csv-benchmark/yellow_tripdata_2010-01.csv", infer_schema_length=0, n_rows=int(1e7), ignore_errors=True)
%%time
>>> with pl.StringCache():
...    df.select(pl.all().cast(pl.Categorical))
```

# Timings
## master
```
CPU times: user 11.4 s, sys: 515 ms, total: 11.9 s
Wall time: 11.9 s
```

## one may lock/ other low contention
```
CPU times: user 13.4 s, sys: 655 ms, total: 14 s
Wall time: 5.97 s
```

## only low contention builders
```
CPU times: user 17 s, sys: 853 ms, total: 17.8 s
Wall time: 3.32 s
```

## don't builder rev-map under contention
```
CPU times: user 16.6 s, sys: 943 ms, total: 17.5 s
Wall time: 3.01 s
```

## split global_to_local and update_cats over threads
```
CPU times: user 16.2 s, sys: 917 ms, total: 17.1 s
Wall time: 2.93 s
```


## local builders
```
CPU times: user 14.7 s, sys: 646 ms, total: 15.4 s
Wall time: 2.36 s
```